feat: add likes handler and card delete handler

+++ src/utils/api.js
@@ -52,20 +52,20 @@ class Api {
     .then(res => this._checkResStatus(res));
   }
 
-  addLike(cardId){
-    return fetch(`${this._baseURL}/cards/likes/${cardId}`, {
-      method: 'PUT',
-      headers: this._headers
-    })
-    .then(res => this._checkResStatus(res));
-  }
-
-  deleteLike(cardId, card){
-    return fetch(`${this._baseURL}/cards/likes/${cardId}`, {
-      method: 'DELETE',
-      headers: this._headers
-    })
-    .then(res => this._checkResStatus(res));
+  changeLikeCardStatus(cardId, isLiked){
+    if(isLiked){
+      return fetch(`${this._baseURL}/cards/likes/${cardId}`, {
+        method: 'DELETE',
+        headers: this._headers
+      })
+      .then(res => this._checkResStatus(res));
+    } else {
+      return fetch(`${this._baseURL}/cards/likes/${cardId}`, {
+        method: 'PUT',
+        headers: this._headers
+      })
+      .then(res => this._checkResStatus(res));
+    }
   }
 
   deleteCard(cardId){
